refactor: split llm crate facade
